#6 Extrai leitura de diretorio para proprio modulo
